refactor(mapgen): collapse natural wonder planning
