+++ mods/mod-swooper-maps/AGENTS.md
@@ -22,11 +22,12 @@ Scope: `mods/mod-swooper-maps/**`
 - Prefer regenerating `mod/` through the Nx build target over editing build
   artifacts.
 - Placement composes `wonders`, `regions`, and `starts` modules in causal
-  order. Wonders owns map-size demand followed by natural-wonder site planning;
-  regions partitions the playable surface around admitted landmass structure;
-  starts then consumes wonder and region evidence during start planning.
-  Placement recipe steps consume those module ops through the root contract or
-  executable router rather than a flat operation registry.
+  order. Wonders owns pure natural-wonder site planning; the placement recipe
+  admits active Civ7 map-size metadata and passes normalized demand into that
+  planner. Regions partitions the playable surface around admitted landmass
+  structure; starts then consumes wonder and region evidence during start
+  planning. Placement recipe steps consume those module ops through the root
+  contract or executable router rather than a flat operation registry.
 
 ## Ecology domain
