--- mods/mod-swooper-maps/src/domain/placement/modules/wonders/ops/plan-wonders/strategies/index.ts
@@ -1,4 +1,0 @@
-import mapMetadata from "./map-metadata/index.js";
-
-/** Complete executable strategy tuple for map-metadata wonder targets. */
-export default [mapMetadata] as const;
