+++ mods/mod-swooper-maps/src/maps/configs/latest-juicy.config.json
@@ -1238,11 +1238,7 @@
           "config": {}
         }
       },
-      "derive-placement-inputs": {
-        "wonders": {
-          "strategy": "map-metadata",
-          "config": {}
-        },
+      "plan-natural-wonders": {
         "naturalWonders": {
           "strategy": "suitability-diversity",
           "config": {
