+++ mods/mod-swooper-maps/src/maps/configs/mountain-patch.config.json
@@ -1205,11 +1205,7 @@
           "config": {}
         }
       },
-      "derive-placement-inputs": {
-        "wonders": {
-          "strategy": "map-metadata",
-          "config": {}
-        },
+      "plan-natural-wonders": {
         "naturalWonders": {
           "strategy": "suitability-diversity",
           "config": {
