+++ mods/mod-swooper-maps/src/maps/configs/swooper-desert-mountains.config.json
@@ -1205,11 +1205,7 @@
           "config": {}
         }
       },
-      "derive-placement-inputs": {
-        "wonders": {
-          "strategy": "map-metadata",
-          "config": {}
-        },
+      "plan-natural-wonders": {
         "naturalWonders": {
           "strategy": "suitability-diversity",
           "config": {
