+++ mods/mod-swooper-maps/src/recipes/standard/contract-manifest.ts
@@ -48,11 +48,11 @@ import { config as computeShelfConfig } from "./stages/morphology/shelf/steps/co
 import { config as adjustResourcesConfig } from "./stages/placement/steps/adjust-resources/config.js";
 import { config as assignAdvancedStartsConfig } from "./stages/placement/steps/assign-advanced-starts/config.js";
 import { config as assignStartsConfig } from "./stages/placement/steps/assign-starts/config.js";
-import { config as derivePlacementInputsConfig } from "./stages/placement/steps/derive-placement-inputs/config.js";
 import { config as placeDiscoveriesConfig } from "./stages/placement/steps/place-discoveries/config.js";
 import { config as placeNaturalWondersConfig } from "./stages/placement/steps/place-natural-wonders/config.js";
 import { config as placeResourcesConfig } from "./stages/placement/steps/place-resources/config.js";
 import { config as placementConfig } from "./stages/placement/steps/placement/config.js";
+import { config as planNaturalWondersConfig } from "./stages/placement/steps/plan-natural-wonders/config.js";
 import { config as planResourcesConfig } from "./stages/placement/steps/plan-resources/config.js";
 import { config as plotLandmassRegionsConfig } from "./stages/placement/steps/plot-landmass-regions/config.js";
 import { config as preparePlacementSurfaceConfig } from "./stages/placement/steps/prepare-placement-surface/config.js";
@@ -107,7 +107,7 @@ export const standardStageContractManifest = [
   ]),
   stage("map-ecology", [plotBiomesConfig, featuresApplyConfig, plotEffectsConfig]),
   stage("placement", [
-    derivePlacementInputsConfig,
+    planNaturalWondersConfig,
     plotLandmassRegionsConfig,
     placeNaturalWondersConfig,
     preparePlacementSurfaceConfig,
