+++ mods/mod-swooper-maps/src/recipes/standard/stages/placement/index.ts
@@ -3,11 +3,11 @@ import { orderStandardStageSteps } from "../../contract-manifest.js";
 import { AdjustResourcesStep } from "./steps/adjust-resources/step.js";
 import { AssignAdvancedStartsStep } from "./steps/assign-advanced-starts/step.js";
 import { AssignStartsStep } from "./steps/assign-starts/step.js";
-import { DerivePlacementInputsStep } from "./steps/derive-placement-inputs/step.js";
 import { PlaceDiscoveriesStep } from "./steps/place-discoveries/step.js";
 import { PlaceNaturalWondersStep } from "./steps/place-natural-wonders/step.js";
 import { PlaceResourcesStep } from "./steps/place-resources/step.js";
 import { PlacementStep } from "./steps/placement/step.js";
+import { PlanNaturalWondersStep } from "./steps/plan-natural-wonders/step.js";
 import { PlanResourcesStep } from "./steps/plan-resources/step.js";
 import { PlotLandmassRegionsStep } from "./steps/plot-landmass-regions/step.js";
 import { PreparePlacementSurfaceStep } from "./steps/prepare-placement-surface/step.js";
@@ -25,7 +25,7 @@ import { PreparePlacementSurfaceStep } from "./steps/prepare-placement-surface/s
 export default createStage({
   id: "placement",
   steps: orderStandardStageSteps("placement", {
-    "derive-placement-inputs": DerivePlacementInputsStep,
+    "plan-natural-wonders": PlanNaturalWondersStep,
     "plot-landmass-regions": PlotLandmassRegionsStep,
     "place-natural-wonders": PlaceNaturalWondersStep,
     "prepare-placement-surface": PreparePlacementSurfaceStep,
